refactor: Placed each Lambda function into individual folders
