+++ backend/src/lambdas/day_lambda/day_lambda.py
