+++ backend/src/lambdas/exercise_lambda/exercise_lambda.py
