+++ backend/tests/lambdas/test_block_lambda.py
@@ -4,7 +4,7 @@ from unittest.mock import patch, MagicMock
 from tests.base_test import BaseTest
 
 with patch("boto3.resource"):
-    from src.lambdas import block_lambda
+    from src.lambdas.block_lambda import block_lambda
 
 
 class TestBlockLambda(BaseTest):
