+++ backend/tests/lambdas/test_relationship_lambda.py
@@ -4,7 +4,7 @@ from unittest.mock import patch, MagicMock
 from tests.base_test import BaseTest
 
 with patch("boto3.resource"):
-    from src.lambdas import relationship_lambda
+    from src.lambdas.relationship_lambda import relationship_lambda
 
 
 class TestRelationshipLambda(BaseTest):
